% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_healthcheck.R
\name{add_healthcheck}
\alias{add_healthcheck}
\title{Add a Health Check to a Shiny App}
\usage{
add_healthcheck(app, path = "/health")
}
\arguments{
\item{app}{The Shiny app object.}

\item{path}{The URL path for the health check endpoint (default is \verb{/health}).}
}
\value{
The modified \code{shiny.appobj} object with the updated internal \code{httpHandler}
function for handling health check requests.
}
\description{
Adds a health check endpoint to a Shiny app.
}
\examples{
\dontrun{
myapp <- shiny::shinyApp(ui = shiny::fluidPage("hi"), server = function(input, output, session) {})
myapp |> add_healthcheck()
}
}
