% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvFileInput.R
\name{csvFile}
\alias{csvFile}
\title{csvFile: Module server function}
\usage{
csvFile(input, output, session, stringsAsFactors = F)
}
\arguments{
\item{input}{The namespace for the module}

\item{output}{datatable}

\item{session}{session}

\item{stringsAsFactors}{Convert all character columns to factors?}
}
\value{
data.table (data.table package)
}
\description{
Shiny module for csv file input
}
\details{
DETAILS
}
\examples{
library(shiny)
# Only run examples in interactive R sessions

if(interactive()){
  
  ui <- fluidPage(
    sidebarLayout(
      sidebarPanel(
        csvFileInput("datafile", "User data (.csv format)")
      ),
      mainPanel(
        DT::dataTableOutput("table")
      )
    )
  )
  server <- function(input, output, session) {
    datafile <- callModule(csvFile, "datafile",
                           stringsAsFactors = FALSE)
    
    output$table <- DT::renderDataTable({
      datafile()
    })
  }
  
  shinyApp(ui, server)
    
}
}
