% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{.shared_params}
\alias{.shared_params}
\title{Shared Package Parameters}
\arguments{
\item{ui}{User Interface object for the shiny application. Should be a valid \code{shiny.tag.list} or \code{bslib_page} by default.}

\item{server}{Server function for the shiny application or module. Should be a valid R function that takes
\code{input}, \code{output}, and \code{session} as arguments.}

\item{input, output, session}{Default parameters for shiny server functions.}

\item{app}{A shiny application object, typically created using \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}

\item{req}{A shiny request object}

\item{resp}{A shiny response object}
}
\description{
Common, shared parameters that can be inherited by other functions in the package.

Use \verb{@inheritParams .shared_params} in a function's roxygen2 block to import these parameter descriptions.
}
\keyword{internal}
