% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{.shared_returns_check}
\alias{.shared_returns_check}
\title{Check Returns}
\value{
If the checks pass, invisibly returns the provided object, \code{x}. If checks fail, a condition error of class \code{check_error}
is thrown.
}
\description{
Returns for the \verb{check_*()} functions.
Use \verb{@inherit .shared_returns_check returns} in a function's roxygen2 block to apply this return description.
}
\keyword{internal}
