% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyhealth-package.R
\docType{package}
\name{shinyhealth-package}
\alias{shinyhealth}
\alias{shinyhealth-package}
\title{\code{shinyhealth} Package}
\description{
The \code{shinyhealth} package provides tools for incorporating a simple
health check (\verb{/health}) endpoint for R Shiny applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig/shinyhealth}
  \item \url{http://docs.jimbrig.com/shinyhealth/}
  \item Report bugs at \url{https://github.com/jimbrig/shinyhealth/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

Authors:
\itemize{
  \item Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})
}

}
\keyword{internal}
